feat: async batch message previews
These changes make the fetching of the latest conversation messages an
async background task that fetches messages in batches of 50.

This will allow the conversations list to render immediately after
profiles are fetched. During the batched requests, a loading indicator
is displayed in the selected tab.

Until we implement message preview caching, the user experience during
the loading of message previews will degrade as the previews stream in.
This will likely cause some conversations to jump around from sorting or
disappear completely because the conversation has no messages.

Even with proper caching, the initial loading of conversations and their
preview messages will take time; however, this should only occur when
initially signing into an account on a new browser/device.

In the future, we can mitigate this initial degraded user experience by
exporting and then importing conversation and message data.