feat: :sparkles: Added cache invalidation/reset on client side
This pull request clears the cache after a logout.  The purpose of this PR is to prevent page refreshes on disconnection:   - On disconnection, the user profile stored in the `useProfileStore` is deleted (set to `null`) , along with the `accessToken`, allowing the UI to update itself and future requests to no longer be authenticated in order to obtain neutral data.   - The Apollo client cache is also reset, and the reset event is also intercepted (using a custom hook called `useResetStoreEvent`) in various components to update their UI and/or refresh data.